bot-telegram/gateway-py/requirements.lock.txt
# Production lock for Linux VPS deploys.
anyio==4.4.0
certifi==2024.8.30
exceptiongroup==1.2.2 ; python_version < "3.11"
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==21.10
sniffio==1.3.1
typing-extensions==4.12.2
